Refactor: threeStyleをAPIから引く部分を抽出
